+++ extensions/broker-server-api/api/src/main/java/de/sovity/edc/ext/brokerserver/api/BrokerServerResource.java
@@ -14,6 +14,7 @@
 
 package de.sovity.edc.ext.brokerserver.api;
 
+import de.sovity.edc.ext.brokerserver.api.model.ConnectorCreationRequest;
 import de.sovity.edc.ext.brokerserver.api.model.AuthorityPortalConnectorDataOfferInfo;
 import de.sovity.edc.ext.brokerserver.api.model.AuthorityPortalConnectorInfo;
 import de.sovity.edc.ext.brokerserver.api.model.AuthorityPortalOrganizationMetadataRequest;
@@ -75,6 +76,12 @@ public interface BrokerServerResource {
     @Consumes(MediaType.APPLICATION_JSON)
     @Operation(description = "Add unknown Connectors to the Broker Server")
     void addConnectors(List<String> endpoints, @QueryParam("adminApiKey") String adminApiKey);
+
+    @PUT
+    @Path("connectors-with-mdsid")
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Operation(description = "Add unknown Connectors with MDS IDs to the Broker Server")
+    void addConnectorsWithMdsIds(ConnectorCreationRequest connectors, @QueryParam("adminApiKey") String adminApiKey);
 
     @DELETE
     @Path("connectors")